rrr: convert misc/stat and misc/netinfo to modules
Two header-only conversions batched. Both had a single includer
(rrr.hpp) so the consumer churn is minimal.

  rrr.stat        — AvgStat POD with sample/clear/reset/peek/avg.
  rrr.netinfo     — NetInfo singleton reading
                    /sys/class/net/ens4/statistics/{rx,tx}_bytes.
                    Uses std::ifstream + std::getline (visible via
                    `import std;`).

Both required `#include <cstdint>` in GMF for the type aliases
(int64_t, unsigned long is fine via Cstdlib).

Measurement (rrr + rpcbench, -j32, clean):
  +internal_protocol:  38.66 s, 175.3 MB PCM
  +stat+netinfo:       38.23 s, 175.4 MB PCM

14 modules total; wallclock essentially flat. PCM creep is small for
header-only conversions (~100 KB per module after dedup).
